Refactoring: Extracted Order class
