+++ spec/order_spec.rb
@@ -1,0 +1,12 @@
+require 'order'
+
+describe Order do
+  subject(:order) { described_class.new }
+  let(:menu) { double() }
+
+  it 'allows select dishes and quantities' do
+    order.select_dish("Chicken Curry", 2)
+    expect(order.dishes).to include({:name => "Chicken Curry", :quantity => 2})
+  end
+
+end
